futex_lock_pi and futex_unlock_pi, sort of

+++ waterbox/waterboxhost/src/context/mod.rs
@@ -63,7 +63,7 @@ pub type SyscallCallback = extern "sysv64" fn(
 /// Layout must be synced with interop.s
 pub struct Context {
 	/// thread id.  1 is main thread and has different call procedures
-	pub tid: usize,
+	pub tid: u32,
 	/// thread pointer as set by guest libc (pthread_self, etc)
 	pub thread_area: usize,
 	/// a lock that this thread is waiting on
@@ -105,7 +105,7 @@ pub struct ContextCallInfo {
 
 impl Context {
 	/// Returns a suitably initialized context.  It's almost ready to use, but host_ptr must be set before each usage
-	pub fn new(tid: usize, context_call_info: *const ContextCallInfo, initial_guest_rsp: usize) -> Context {
+	pub fn new(tid: u32, context_call_info: *const ContextCallInfo, initial_guest_rsp: usize) -> Context {
 		let mut res: Context = unsafe { std::mem::zeroed() };
 		res.tid = tid;
 		res.context_call_info = context_call_info;
